(Java) Install all of the necessary Java files
We need to install all of the Java build products
in $ROSE[_BUILD]/src/3rdPartyLibraries/java-parser
otherwise ROSE won't be able to find our classes.

Particularly, `jserver_FindClass("JavaTraversal")`
returns NULL because the class is not loaded in
the JVM within the ECJ-ROSE Connection.

The installed ROSE translator uses this classpath:

    -Djava.class.path=
    $ROSE_INSTALL/lib/antlr-3.3-complete.jar
    $ROSE_INSTALL/lib/OpenFortranParser-20111001.jar
    $ROSE_INSTALL/lib/ecjROSE.jar
    $ROSE_INSTALL/lib/

Before this patch, `ecjROSE.jar` and all of the
`java-parser/*class` files were missing.
Particularly, `JavaTraversal.class` was missing.
Installing all of these files resolves this issue.

Also, as a precautionary measure, we add some
assertions in ecj.C to ensure that the
"JavaTraversal" class is found successfully. An
assertion is a lot more friendly and useful than
a JVM crash report.