(CLI) Add -rose:keep_going
This commandline option is similar to GNU Make's --keep-going
option, except it will never fail (assuming the input code
is valid).

The point of this option to provide a failover mechanism for
ROSE to completely compile large applications, even if some
of the files fail within the internal processings of ROSE.

If any internal ROSE errors occur, ROSE will simply pass the
original input file to the backend compiler. This also includes
the case where ROSE processes a file without failure, but the
unparsed file is actually invalid.

In order to use this option, ROSE must be configured to disable
assertions:

  $ CXXFLAGS="-DNDEBUG" $ROSE/configure ...

With assertions disabled, ROSE can run continuously, rather than
halting or aborting, and thereby terminating the program execution.
Instead, we throw and catch exceptions, or detect hardware signals
like SIGABRT (abort) and SIGSEGV (segmentation fault).